:broom: :spider_web: Fix syntax mistakes caught by Markdown linter
This commit used a Markdown linter to catch several warnings and errors
with Markdown syntax. This commit fixes most of the warnings that come
out of the linter. This Pull Request will also trigger the forthcoming
Circle CI change to pass; on first implementation, it will be broken
until this commit is merged.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>