:boom: Delete old HTML build artifacts no longer in use
This commit deletes all the HTML-generated files for the website in the
`build/` directory. We want this directory to be cleared when the Circle
CI job goes to rebuild the website, and we do not want to commit those
files into this repository.

I split this out into a new Pull Request since it changes so many files,
and I did not want to make #3 or #4 too lengthy to review.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>